requirements.txt
pyrogram>=2.0.0
aiosqlite
python-dotenv
